backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pandas==2.2.3
networkx==3.4.2
google-generativeai==0.8.3
python-dotenv==1.0.1
pydantic==2.10.3
pytest==9.0.2
httpx==0.28.1
